versioning tests moved to lib/versioning and fixed the path filter issuesarrised due to changes
