+++ lib/versioning/hashicorp/hashicorp.spec.ts
@@ -1,4 +1,4 @@
-import { api as semver } from '../../lib/versioning/hashicorp';
+import { api as semver } from '.';
 
 describe('semver.matches()', () => {
   it('handles tilde greater than', () => {
