+++ lib/versioning/index.spec.ts
@@ -1,10 +1,7 @@
-import * as allVersioning from '../../lib/versioning';
-import { getOptions } from '../../lib/config/definitions';
-import {
-  GenericVersioningApi,
-  GenericVersion,
-} from '../../lib/versioning/loose/generic';
-import * as semverVersioning from '../../lib/versioning/semver';
+import * as allVersioning from '.';
+import { getOptions } from '../config/definitions';
+import { GenericVersioningApi, GenericVersion } from './loose/generic';
+import * as semverVersioning from './semver';
 
 const supportedSchemes = getOptions().find(
   option => option.name === 'versioning'
