+++ lib/versioning/npm/npm.spec.ts
@@ -1,4 +1,4 @@
-import { api as semver } from '../../lib/versioning/npm';
+import { api as semver } from '.';
 
 describe('semver.isValid(input)', () => {
   it('should return null for irregular versions', () => {
